Remove mod_deploy module, update CD Github workflow
