+++ config_sample.py
@@ -13,7 +13,6 @@ SERVER_NAME = 'localhost'
 EMAIL_DOMAIN = ''
 EMAIL_API_KEY = ''
 HMAC_KEY = ''
-GITHUB_DEPLOY_KEY = ''
 GITHUB_CI_KEY = ''
 GITHUB_CLIENT_ID = ''
 GITHUB_CLIENT_KEY = ''
