--- mod_deploy/__init__.py
@@ -1,1 +1,0 @@
-"""contains methods and classes for GitHub hooks based on auto deployment."""
