--- tests/test_deploy/__init__.py
@@ -1,1 +1,0 @@
-"""Contains tests for mod_deploy."""
